Create Cloud Caluclator
This application is a simple command-line program that uses the Weatherstack API to get the current weather conditions for a specific location and checks if there are any bad weather clouds approaching the location.

The program first initializes libcurl, a library for making HTTP requests, and sets options such as the URL of the API endpoint and the function to handle the HTTP response. It then performs the HTTP request and stores the response in a string.

The response, which is in JSON format, is parsed using the nlohmann/json library. The program checks for any errors in the response and proceeds to check if there are any bad weather clouds approaching the location. In this case, a bad weather cloud is defined as a cloud with a cloud cover percentage of over 50%.

If a bad weather cloud is found, the program outputs a message indicating the type of cloud and the approximate time when it will start raining. The observation time of the weather conditions is also included in the output. If no bad weather clouds are found, the program outputs a message indicating that there is no bad weather approaching the location.

The program can be modified to include additional weather information, such as temperature, humidity, and UV index, by modifying the API request URL to include the desired fields in the response.